+++ CHANGELOG.md
@@ -15,6 +15,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ## [Unreleased]
 ### Added
 - InvoicePullSubscription and InvoicePullRequest resources
+### Changed
+- create method from Transaction resource. This method is now deprecated
 
 ## [2.19.0] - 2025-03-17
 ### Added